New `Tuning` struct to clean up tuning related code; fix for `keydownduration` being big-endian and not little-endian; fixed `ftune` conversion based on new information (see note below); reworked the `apply_zone_data_to_split` calls to be more accurate to the sf2 specs
Note: The 4th `sample_rate_adjustment_curve` has been removed, with the default now set to the 1st curve (ideal curve). The reason the 100 cents shift was necessary was because the `ftune` parameter is not actually a signed integer representing a shift in pitch in cents, but instead an unsigned integer representing an upwards-shift in pitch mapped accordingly: 255 = ~50 cents of upward pitch shifting. Since the old implementation of the tuning parameter converter didn't know this, it would directly assign the signed cents value to the `ftune` parameter, and the 100 cents shift downwards in the sample rate adjustment curve was necessary to kind of "average out" the discrepency caused by negative shifts like -127 cents being interpreted by the game as a ~50 cents upward shift since -127 is 0xFF in two's complement.